# docs/api/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "NovaRewards API",
    "version": "1.0.0",
    "description": "REST API for the Nova Rewards blockchain loyalty platform. User endpoints require a Bearer JWT from `POST /auth/login`. Merchant endpoints require an `x-api-key` header.",
    "contact": {
      "name": "Nova Rewards Team"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local development"
    },
    {
      "url": "https://api.novarewards.io/api",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT access token from `POST /auth/login`."
      },
      "merchantApiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Merchant API key returned on registration (shown once)."
      }
    },
...
  },
  "paths": {
...
    "/rewards/issue": {
      "post": {
        "tags": [
          "Rewards"
        ],
        "summary": "Enqueue idempotent reward issuance",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idempotencyKey",
                  "walletAddress",
                  "amount",
                  "campaignId"
                ],
                "properties": {
                  "idempotencyKey": {
                    "type": "string"
                  },
                  "walletAddress": {
                    "type": "string"
                  },
                  "amount": {
                    "type": "number"
                  },
                  "campaignId": {
                    "type": "integer"
                  },
                  "userId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Duplicate — already processed"
          },
          "202": {
            "description": "Job queued"
          },
          "400": {
            "description": "Validation error"
          }
        }
      }
    },
    "/rewards/distribute": {
      "post": {
        "tags": [
          "Rewards"
        ],
        "summary": "Distribute NOVA tokens to a customer wallet",
        "security": [
          {
            "merchantApiKey": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "walletAddress",
                  "amount"
                ],
                "properties": {
                  "walletAddress": {
                    "type": "string",
                    "example": "GBBD47..."
                  },
                  "amount": {
                    "type": "number",
                    "example": 50
                  },
                  "campaignId": {
                    "type": "integer",
                    "example": 3
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tokens distributed",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "txHash": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "No trustline or validation error"
          },
          "403": {
            "description": "Campaign not owned by merchant"
          },
          "404": {
            "description": "Campaign not found"
          },
          "502": {
            "description": "Chain error"
          }
        }
      }
    },
    "/redemptions": {
      "post": {
        "tags": [
          "Redemptions"
        ],
        "summary": "Redeem a reward (idempotent)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "X-Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "userId",
                  "rewardId"
                ],
                "properties": {
                  "userId": {
                    "type": "integer"
                  },
                  "rewardId": {
                    "type": "integer"
                  },
                  "campaignId": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Idempotent replay"
          },
          "201": {
            "description": "Redemption created"
          },
          "400": {
            "description": "Validation error"
          },
          "403": {
            "description": "Redeeming for another user"
          },
          "404": {
            "description": "Reward not found"
          },
          "409": {
            "description": "Out of stock / insufficient points / inactive"
          }
        }
      },
      "get": {
        "tags": [
          "Redemptions"
        ],
        "summary": "List redemption history",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated redemptions"
          },
          "401": {
            "description": "Unauthenticated"
          }
        }
      }
    },
    "/redemptions/{id}": {
      "get": {
        "tags": [
          "Redemptions"
        ],
        "summary": "Get redemption by ID",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Redemption record"
          },
          "400": {
            "description": "Invalid id"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
  }
}